Repo Organization
Put items into folders for better organization. Added updated code for arm control called "ArmControl.py".
