+++ Controller Implementation/Controller Mapping.txt
@@ -1,0 +1,30 @@
+DUALSHOCK4 ARM CONTROLS
+------------------------------------------------------------------------------------------------
+
+STAGE 1: Calibration Controls (No constraints set until X press):
+------------------------------------------------------------------------------------------------
+
+D-Pad Left ---> Set zero position
+
+D-Pad Up -----> Return to zero position
+
+SHARE --------> Set calibration and move into stage 2
+
+
+STAGE 2: Live Arm Movement:
+------------------------------------------------------------------------------------------------
+L2, R2 -------> Rotate the base
+
+LUp, LDown ---> Rotate shoulder (controls refer to Ref motor, other motor is inversed)
+
+RUp, RDown ---> Rotate elbow
+
+L1, R1 -------> Rotate wrist
+
+X ------------> Close claw
+
+∆ ------------> Open claw
+
+O (Hold) -----> While O is held, disable constraints
+
+OPTIONS ------> Return to calibration mode
