--- Electrical Components for Arm.txt
@@ -1,34 +1,0 @@
--------------------- VINT HUB --------------------------
-Vint Hub (2):
-https://www.phidgets.com/?tier=3&catid=2&pcid=1&prodid=643
--> Interface between Raspberry PI (VIA USB) and motor drivers
-
---------------- MOTOR/MOTOR DRIVERS ---------------------------
-Motor Drivers (6, 1 per motor):
-https://www.phidgets.com/?tier=3&catid=23&pcid=20&prodid=1119
--> One motor driver connected to each motor
-
-Nema-23 Motor 77:1 (1):
-https://www.phidgets.com/?tier=3&catid=24&pcid=21&prodid=353
--> For the base rotation
-
-Nema-23 Motor 15:1 (3):
-https://www.phidgets.com/?tier=3&catid=24&pcid=21&prodid=352
--> 2 for bending the shoulder and 1 for bending the elbow
-
-Nema-11 Motor 100:1 (1):
-https://www.phidgets.com/?tier=3&catid=24&pcid=21&prodid=341
--> For rotating the claw
-
-Nema-11 Gearless Motor (1):
-https://www.phidgets.com/?tier=3&catid=24&pcid=21&prodid=339
--> For pinching the claw
-
---------------- ENCODERS/ENCODER DRIVERS -------------------------
-Quadrature Encoder (6, 1 per encoder):
-https://www.phidgets.com/?tier=3&catid=4&pcid=2&prodid=959
--> One quadrature encoder connected to each encoder
-
-Optical Rotary Encoder (6, 1 per motor):
-https://www.phidgets.com/?tier=3&catid=103&pcid=83&prodid=404
--> Measures rotation of motors even for missteps
